+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-v2.0.0 (in development)
------------------------
+v2.0.0 (2025-12-19)
+-------------------
 - Support Python 3.14
 - Drop support for Python 3.8 and 3.9
 - Changed `ParsedWheelFilename` from a namedtuple to a dataclass